backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
python-jose
gunicorn
python-dotenv
requests
flask-cors
psycopg2